+++ Polyhedron/demo/Polyhedron/Polyhedron_demo_detect_sharp_edges.h
@@ -2,7 +2,7 @@
 #ifndef POLYHEDRON_DEMO_DETECT_SHARP_EDGES_H
 #define POLYHEDRON_DEMO_DETECT_SHARP_EDGES_H
 
-#include <CGAL/Polygon_mesh_processing/Detect_features_in_polyhedra.h>
+#include <CGAL/Polygon_mesh_processing/internal/Detect_features_in_polyhedra.h>
 #include <CGAL/Mesh_3/properties.h>
 
 namespace CGAL
